pyproject.toml
[project]
name = "eval-menu"
version = "0.1.0"
description = "Évaluation en terme d'impact environnemental et santé d’un menu."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.3",
    "djangorestframework",
]
